autoformat all code using ocamlformat 0.9
This uses the `conventional` profile to minimise surprise,
and brings the style inline to the autoformatting conventions
being adopted by the OCaml community.

Since datakit is essentially stable now, I also took the
liberty of bumping the release version of this to 1.0.0
so that it can be parked.

+++ src/datakit-io/datakit_io.ml
@@ -1,35 +1,39 @@
 open Lwt.Infix
 
 let src = Logs.Src.create "irmin-io" ~doc:"Datakit sync support"
+
 module Log = (val Logs.src_log src : Logs.LOG)
 
 module IO = struct
   type ctx = unit
+
   let ctx () = Lwt.return (Some ())
+
   type ic = Lwt_io.input_channel
+
   type oc = Lwt_io.output_channel
+
   let write oc s = Lwt_io.write oc s
+
   let flush oc = Lwt_io.flush oc
 
   let with_ssh_process ?init uri fn =
-    let host = match Uri.host uri with
-      | None   -> "localhost"
-      | Some x -> x
-    in
-    let user = match Uri.userinfo uri with
-      | None   -> ""
-      | Some u -> u ^ "@"
-    in
-    let cmd = match init with
-      | None   -> [| "ssh"; user ^ host; |]
+    let host = match Uri.host uri with None -> "localhost" | Some x -> x in
+    let user = match Uri.userinfo uri with None -> "" | Some u -> u ^ "@" in
+    let cmd =
+      match init with
+      | None -> [| "ssh"; user ^ host |]
       | Some x -> [| "ssh"; user ^ host; x |]
     in
-    Log.info (fun f -> f "Executing '%s'" (String.concat " " (Array.to_list cmd)));
+    Log.info (fun f ->
+        f "Executing '%s'" (String.concat " " (Array.to_list cmd)) );
     let env = Unix.environment () in
     let p = Lwt_process.open_process_full ~env ("ssh", cmd) in
     Lwt.finalize
       (fun () -> fn (p#stdout, p#stdin))
-      (fun () -> let _ = p#close in Lwt.return_unit)
+      (fun () ->
+        let _ = p#close in
+        Lwt.return_unit )
 
   let with_conduit ?init uri fn =
     Log.debug (fun f -> f "Connecting to %s" (Uri.to_string uri));
@@ -40,29 +44,24 @@ module IO = struct
     Conduit_lwt_unix.connect ~ctx client >>= fun (_flow, ic, oc) ->
     Lwt.finalize
       (fun () ->
-         begin match init with
-           | None   -> Lwt.return_unit
-           | Some s -> write oc s
-         end >>= fun () ->
-         fn (ic, oc))
-      (fun ()  ->
-         Lwt.catch
-           (fun () -> Lwt_io.close ic)
-           (function
-             | Unix.Unix_error _ -> Lwt.return_unit
-             | e -> Lwt.fail e))
+        (match init with None -> Lwt.return_unit | Some s -> write oc s)
+        >>= fun () -> fn (ic, oc) )
+      (fun () ->
+        Lwt.catch
+          (fun () -> Lwt_io.close ic)
+          (function Unix.Unix_error _ -> Lwt.return_unit | e -> Lwt.fail e) )
 
   let with_connection ?ctx:_ uri ?init fn =
     match Git.Sync.protocol uri with
     | `Ok `SSH -> with_ssh_process ?init uri fn
     | `Ok `Git -> with_conduit ?init uri fn
     | `Ok `Smart_HTTP ->
-      (* HTTP.with_http ?init (with_conduit ?init:None) uri fn *)
-      Lwt.fail (Failure ("smart HTTP is not supported. Use git:// or SSH"))
+        (* HTTP.with_http ?init (with_conduit ?init:None) uri fn *)
+        Lwt.fail (Failure "smart HTTP is not supported. Use git:// or SSH")
     | `Not_supported x ->
-      Lwt.fail (Failure ("Scheme " ^ x ^ " not supported yet"))
+        Lwt.fail (Failure ("Scheme " ^ x ^ " not supported yet"))
     | `Unknown ->
-      Lwt.fail (Failure ("Unknown protocol. Must supply a scheme like git://"))
+        Lwt.fail (Failure "Unknown protocol. Must supply a scheme like git://")
 
   let read_all ic =
     let len = 4 * 4096 in
@@ -72,9 +71,8 @@ module IO = struct
       Lwt_io.read_into ic buf 0 len >>= function
       | 0 -> return acc
       | i ->
-        let buf = Bytes.sub buf 0 i |> Bytes.unsafe_to_string in
-        if len = i then return (buf :: acc)
-        else aux (buf :: acc)
+          let buf = Bytes.sub buf 0 i |> Bytes.unsafe_to_string in
+          if len = i then return (buf :: acc) else aux (buf :: acc)
     in
     aux []
 
@@ -82,14 +80,13 @@ module IO = struct
     let res = Bytes.create n in
     Lwt_io.read_into_exactly ic res 0 n >>= fun () ->
     Lwt.return (Bytes.unsafe_to_string res)
-
 end
 
 module FS = struct
-
   (* From mirage/ocaml-git/lib/unix/git_unix.ml (v1.7.3) *)
 
   let mmap_threshold = 4096
+
   (* Files smaller than this are loaded using [read].
 
      Use of mmap is necessary to handle packfiles efficiently. Since these
@@ -104,6 +101,7 @@ module FS = struct
   *)
 
   let mkdir_pool = Lwt_pool.create 1 (fun () -> Lwt.return_unit)
+
   let openfile_pool = Lwt_pool.create 200 (fun () -> Lwt.return_unit)
 
   let protect_unix_exn = function
@@ -115,46 +113,44 @@ module FS = struct
     | e -> Lwt.fail e
 
   let protect f x = Lwt.catch (fun () -> f x) protect_unix_exn
+
   let safe f x = Lwt.catch (fun () -> f x) ignore_enoent
 
   let mkdir dirname =
     let rec aux dir =
       if Sys.file_exists dir && Sys.is_directory dir then Lwt.return_unit
-      else (
+      else
         let clear =
           if Sys.file_exists dir then (
             Log.debug (fun l ->
-                l "%s already exists but is a file, removing." dir);
-            safe Lwt_unix.unlink dir
-          ) else
-            Lwt.return_unit
+                l "%s already exists but is a file, removing." dir );
+            safe Lwt_unix.unlink dir )
+          else Lwt.return_unit
         in
         clear >>= fun () ->
         aux (Filename.dirname dir) >>= fun () ->
         Log.debug (fun l -> l "mkdir %s" dir);
-        protect (Lwt_unix.mkdir dir) 0o755;
-      ) in
+        protect (Lwt_unix.mkdir dir) 0o755
+    in
     Lwt_pool.use mkdir_pool (fun () -> aux dirname)
 
   let file_exists f = Lwt_unix.file_exists f
 
   module Lock = struct
-
     let is_stale max_age file =
       file_exists file >>= fun exists ->
-      if exists then (
-        Lwt.catch (fun () ->
+      if exists then
+        Lwt.catch
+          (fun () ->
             Lwt_unix.stat file >>= fun s ->
             let stale = Unix.gettimeofday () -. s.Unix.st_mtime > max_age in
-            Lwt.return stale)
+            Lwt.return stale )
           (function
             | Unix.Unix_error (Unix.ENOENT, _, _) -> Lwt.return false
-            | e -> Lwt.fail e)
-      ) else
-        Lwt.return false
+            | e -> Lwt.fail e )
+      else Lwt.return false
 
-    let unlock file =
-      Lwt_unix.unlink file
+    let unlock file = Lwt_unix.unlink file
 
     let lock ?(max_age = 10. *. 60. (* 10 minutes *)) ?(sleep = 0.001) file =
       let rec aux i =
@@ -162,32 +158,35 @@ module FS = struct
         is_stale max_age file >>= fun is_stale ->
         if is_stale then (
           Log.err (fun f -> f "%s is stale, removing it." file);
-          unlock file >>= fun () ->
-          aux 1
-        ) else
+          unlock file >>= fun () -> aux 1 )
+        else
           let create () =
             let pid = Unix.getpid () in
             mkdir (Filename.dirname file) >>= fun () ->
-            Lwt_unix.openfile file [Unix.O_CREAT; Unix.O_RDWR; Unix.O_EXCL] 0o600
+            Lwt_unix.openfile file
+              [ Unix.O_CREAT; Unix.O_RDWR; Unix.O_EXCL ]
+              0o600
             >>= fun fd ->
             let oc = Lwt_io.of_fd ~mode:Lwt_io.Output fd in
-            Lwt_io.write_int oc pid >>= fun () ->
-            Lwt_unix.close fd
+            Lwt_io.write_int oc pid >>= fun () -> Lwt_unix.close fd
           in
           Lwt.catch create (function
-              | Unix.Unix_error(Unix.EEXIST, _, _) ->
-                let backoff = 1. +. Random.float (let i = float i in i *. i) in
-                Lwt_unix.sleep (sleep *. backoff) >>= fun () ->
-                aux (i+1)
-              | e -> Lwt.fail e)
+            | Unix.Unix_error (Unix.EEXIST, _, _) ->
+                let backoff =
+                  1.
+                  +. Random.float
+                       (let i = float i in
+                        i *. i)
+                in
+                Lwt_unix.sleep (sleep *. backoff) >>= fun () -> aux (i + 1)
+            | e -> Lwt.fail e )
       in
       aux 1
 
     let with_lock file fn =
       match file with
-      | None   -> fn ()
+      | None -> fn ()
       | Some f -> lock f >>= fun () -> Lwt.finalize fn (fun () -> unlock f)
-
   end
 
   let mkdir = mkdir
@@ -196,6 +195,7 @@ module FS = struct
 
   (* we use file locking *)
   type lock = path
+
   let lock_file x = x
 
   let file_exists = file_exists
@@ -208,44 +208,43 @@ module FS = struct
       let d = List.filter kind d in
       let d = List.sort String.compare d in
       Lwt.return d
-    else
-      Lwt.return_nil
+    else Lwt.return_nil
 
   let directories dir =
-    list_files (fun f ->
-        try Sys.is_directory f with Sys_error _ -> false
-      ) dir
+    list_files
+      (fun f -> try Sys.is_directory f with Sys_error _ -> false)
+      dir
 
   let files dir =
-    list_files (fun f ->
-        try not (Sys.is_directory f) with Sys_error _ -> false
-      ) dir
+    list_files
+      (fun f -> try not (Sys.is_directory f) with Sys_error _ -> false)
+      dir
 
   let write_cstruct fd b =
     let rec rwrite fd buf ofs len =
       Lwt_bytes.write fd buf ofs len >>= fun n ->
       if len = 0 then Lwt.fail End_of_file
       else if n < len then rwrite fd buf (ofs + n) (len - n)
-      else Lwt.return_unit in
+      else Lwt.return_unit
+    in
     match Cstruct.len b with
-    | 0   -> Lwt.return_unit
+    | 0 -> Lwt.return_unit
     | len -> rwrite fd (Cstruct.to_bigarray b) 0 len
 
-  let delays = Array.init 20 (fun i -> 0.1 *. (float i) ** 2.)
+  let delays = Array.init 20 (fun i -> 0.1 *. (float i ** 2.))
 
   let command fmt =
-    Printf.ksprintf (fun str ->
+    Printf.ksprintf
+      (fun str ->
         Log.debug (fun l -> l "[exec] %s" str);
         let i = Sys.command str in
         if i <> 0 then Log.debug (fun l -> l "[exec] error %d" i);
-        Lwt.return_unit
-      ) fmt
+        Lwt.return_unit )
+      fmt
 
   let remove_dir dir =
-    if Sys.os_type = "Win32" then
-      command "cmd /d /v:off /c rd /s /q %S" dir
-    else
-      command "rm -rf %S" dir
+    if Sys.os_type = "Win32" then command "cmd /d /v:off /c rd /s /q %S" dir
+    else command "rm -rf %S" dir
 
   let remove_file ?lock file =
     Lock.with_lock lock (fun () ->
@@ -256,58 +255,53 @@ module FS = struct
                rename a file or directory or to remove an existing
                directory. *)
             | Unix.Unix_error (Unix.EACCES, _, _)
-            | Unix.Unix_error (Unix.EISDIR, _, _) -> remove_dir file
+            | Unix.Unix_error (Unix.EISDIR, _, _) ->
+                remove_dir file
             | Unix.Unix_error (Unix.ENOENT, _, _) -> Lwt.return_unit
-            | e -> Lwt.fail e)
-      )
+            | e -> Lwt.fail e ) )
 
   let rename =
     if Sys.os_type <> "Win32" then Lwt_unix.rename
-    else
-      fun tmp file ->
-        let rec aux i =
-          Lwt.catch
-            (fun () -> Lwt_unix.rename tmp file)
-            (function
-              (* On Windows, [EACCES] can also occur in an attempt to
+    else fun tmp file ->
+      let rec aux i =
+        Lwt.catch
+          (fun () -> Lwt_unix.rename tmp file)
+          (function
+            (* On Windows, [EACCES] can also occur in an attempt to
                  rename a file or directory or to remove an existing
                  directory. *)
-              | Unix.Unix_error (Unix.EACCES, _, _) as e ->
+            | Unix.Unix_error (Unix.EACCES, _, _) as e ->
                 if i >= Array.length delays then Lwt.fail e
-                else (
+                else
                   file_exists file >>= fun exists ->
-                  if exists && Sys.is_directory file then (
-                    remove_dir file >>= fun () -> aux (i+1)
-                  ) else (
+                  if exists && Sys.is_directory file then
+                    remove_dir file >>= fun () -> aux (i + 1)
+                  else (
                     Log.debug (fun l ->
-                        l "Got EACCES, retrying in %.1fs" delays.(i));
-                    Lwt_unix.sleep delays.(i) >>= fun () -> aux (i+1)
-                  ))
-              | e -> Lwt.fail e)
-        in
-        aux 0
+                        l "Got EACCES, retrying in %.1fs" delays.(i) );
+                    Lwt_unix.sleep delays.(i) >>= fun () -> aux (i + 1) )
+            | e -> Lwt.fail e )
+      in
+      aux 0
 
   let with_write_file ?temp_dir file fn =
-    begin match temp_dir with
-      | None   -> Lwt.return_unit
-      | Some d -> mkdir d
-    end >>= fun () ->
+    (match temp_dir with None -> Lwt.return_unit | Some d -> mkdir d)
+    >>= fun () ->
     let dir = Filename.dirname file in
     mkdir dir >>= fun () ->
     let tmp = Filename.temp_file ?temp_dir (Filename.basename file) "write" in
     Lwt_pool.use openfile_pool (fun () ->
         Log.debug (fun l -> l "Writing %s (%s)" file tmp);
-        Lwt_unix.(openfile tmp [O_WRONLY; O_NONBLOCK; O_CREAT; O_TRUNC] 0o644)
+        Lwt_unix.(
+          openfile tmp [ O_WRONLY; O_NONBLOCK; O_CREAT; O_TRUNC ] 0o644)
         >>= fun fd ->
         Lwt.finalize (fun () -> protect fn fd) (fun () -> Lwt_unix.close fd)
-        >>= fun () ->
-        rename tmp file
-      )
+        >>= fun () -> rename tmp file )
 
   let read_file_with_read file size =
     let chunk_size = max 4096 (min size 0x100000) in
     let buf = Cstruct.create size in
-    let flags = [Unix.O_RDONLY] in
+    let flags = [ Unix.O_RDONLY ] in
     let perm = 0o0 in
     Lwt_unix.openfile file flags perm >>= fun fd ->
     let rec aux off =
@@ -317,34 +311,29 @@ module FS = struct
          real size of the file. This may happen for instance if the
          file was truncated while reading. *)
       let off = off + read in
-      if off >= size then
-        Lwt.return buf
-      else
-        aux off
+      if off >= size then Lwt.return buf else aux off
     in
-    Lwt.finalize (fun () -> aux 0)
-      (fun () -> Lwt_unix.close fd)
+    Lwt.finalize (fun () -> aux 0) (fun () -> Lwt_unix.close fd)
 
   let read_file_with_mmap file =
-    let fd = Unix.(openfile file [O_RDONLY; O_NONBLOCK] 0o644) in
+    let fd = Unix.(openfile file [ O_RDONLY; O_NONBLOCK ] 0o644) in
     let ba = Lwt_bytes.map_file ~fd ~shared:false () in
     Unix.close fd;
     Lwt.return (Cstruct.of_bigarray ba)
 
   let read_file file =
-    Lwt.catch (fun () ->
+    Lwt.catch
+      (fun () ->
         Lwt_pool.use openfile_pool (fun () ->
             Log.debug (fun l -> l "Reading %s" file);
             Lwt_unix.stat file >>= fun stats ->
             let size = stats.Lwt_unix.st_size in
-            (if size >= mmap_threshold then read_file_with_mmap file
-             else read_file_with_read file size
-            ) >|= fun buf ->
-            Some buf
-          )
-      ) (function
-        | Unix.Unix_error _ | Sys_error _ -> Lwt.return_none
-        | e -> Lwt.fail e)
+            ( if size >= mmap_threshold then read_file_with_mmap file
+            else read_file_with_read file size )
+            >|= fun buf -> Some buf ) )
+      (function
+        | Unix.Unix_error _ | Sys_error _ -> Lwt.return_none | e -> Lwt.fail e
+        )
 
   let stat_info_unsafe path =
     let open Git.Index in
@@ -353,25 +342,27 @@ module FS = struct
     let mtime = { lsb32 = Int32.of_float stats.Unix.st_mtime; nsec = 0l } in
     let dev = Int32.of_int stats.Unix.st_dev in
     let inode = Int32.of_int stats.Unix.st_ino in
-    let mode = match stats.Unix.st_kind, stats.Unix.st_perm with
+    let mode =
+      match (stats.Unix.st_kind, stats.Unix.st_perm) with
       | Unix.S_REG, p -> if p land 0o100 = 0o100 then `Exec else `Normal
       | Unix.S_LNK, _ -> `Link
       | k, p ->
-        let kind = match k with
-          | Unix.S_REG -> "REG"
-          | Unix.S_DIR -> "DIR"
-          | Unix.S_CHR -> "CHR"
-          | Unix.S_BLK -> "BLK"
-          | Unix.S_LNK -> "LNK"
-          | Unix.S_FIFO -> "FIFO"
-          | Unix.S_SOCK -> "SOCK"
-        in
-        let perm = Printf.sprintf "%o" p in
-        let error =
-          Printf.sprintf "%s: not supported kind of file [%s, %s]."
-            path kind perm
-        in
-        failwith error
+          let kind =
+            match k with
+            | Unix.S_REG -> "REG"
+            | Unix.S_DIR -> "DIR"
+            | Unix.S_CHR -> "CHR"
+            | Unix.S_BLK -> "BLK"
+            | Unix.S_LNK -> "LNK"
+            | Unix.S_FIFO -> "FIFO"
+            | Unix.S_SOCK -> "SOCK"
+          in
+          let perm = Printf.sprintf "%o" p in
+          let error =
+            Printf.sprintf "%s: not supported kind of file [%s, %s]." path kind
+              perm
+          in
+          failwith error
     in
     let uid = Int32.of_int stats.Unix.st_uid in
     let gid = Int32.of_int stats.Unix.st_gid in
@@ -379,9 +370,11 @@ module FS = struct
     { ctime; mtime; dev; inode; uid; gid; mode; size }
 
   let stat_info path =
-    Lwt.catch (fun () -> Lwt.return (Some (stat_info_unsafe path))) (function
-        | Sys_error _ | Unix.Unix_error _ -> Lwt.return_none
-        | e -> Lwt.fail e)
+    Lwt.catch
+      (fun () -> Lwt.return (Some (stat_info_unsafe path)))
+      (function
+        | Sys_error _ | Unix.Unix_error _ -> Lwt.return_none | e -> Lwt.fail e
+        )
 
   let chmod ?lock f `Exec =
     Lock.with_lock lock (fun () -> Lwt_unix.chmod f 0o755)
@@ -392,38 +385,33 @@ module FS = struct
     in
     Lock.with_lock lock (fun () ->
         Lwt.catch write (function
-            | Unix.Unix_error (Unix.EISDIR, _, _) -> remove_dir file >>= write
-            | e -> Lwt.fail e
-          )
-      )
+          | Unix.Unix_error (Unix.EISDIR, _, _) -> remove_dir file >>= write
+          | e -> Lwt.fail e ) )
 
   let trim s =
     let len = Cstruct.len s in
-    if len >= 2
-    && Cstruct.get_char s (len - 1) = '\n'
-    && Cstruct.get_char s (len - 2) = '\r'
-    then
-      Cstruct.sub s 0 (len - 2)
+    if
+      len >= 2
+      && Cstruct.get_char s (len - 1) = '\n'
+      && Cstruct.get_char s (len - 2) = '\r'
+    then Cstruct.sub s 0 (len - 2)
     else if len >= 1 && Cstruct.get_char s (len - 1) = '\n' then
       Cstruct.sub s 0 (len - 1)
-    else
-      s
+    else s
 
   let test_and_set_file ?temp_dir ~lock file ~test ~set =
     Lock.with_lock (Some lock) (fun () ->
         read_file file >>= fun v ->
-        let equal = match test, v with
-          | None  , None   -> true
+        let equal =
+          match (test, v) with
+          | None, None -> true
           | Some x, Some y -> Cstruct.equal (trim x) (trim y)
           | _ -> false
         in
         if not equal then Lwt.return false
         else
-          (match set with
-           | None   -> remove_file file
-           | Some v -> write_file ?temp_dir file v)
-          >|= fun () ->
-          true
-      )
-
+          ( match set with
+          | None -> remove_file file
+          | Some v -> write_file ?temp_dir file v )
+          >|= fun () -> true )
 end
